src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "ollama>=0.6.1",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "lightrag-hku[api]",
    "neo4j>=5.0.0",
    "qdrant-client>=1.6.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "aiohttp>=3.8.0",
    "json-repair>=0.1.0",
    "tenacity>=8.0.0",
    "composio-core>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live: marks tests that require live external integrations and credentials",
    "fast: marks deterministic tests suitable for local and CI default runs",
]
